vttablet: address review comments

+++ go/vt/vttablet/tabletserver/tabletserver_test.go
@@ -2162,7 +2162,7 @@ func setupTabletServerTest(t *testing.T) (*fakesqldb.DB, *TabletServer) {
 func setupTabletServerTestCustom(t *testing.T, config *tabletenv.TabletConfig) (*fakesqldb.DB, *TabletServer) {
 	db := setupFakeDB(t)
 	tsv := NewTabletServer("TabletServerTest", config, memorytopo.NewServer(""), topodatapb.TabletAlias{})
-	require.Equal(t, int64(StateNotConnected), tsv.sm.State())
+	require.Equal(t, StateNotConnected, tsv.sm.State())
 	dbcfgs := newDBConfigs(db)
 	target := querypb.Target{TabletType: topodatapb.TabletType_MASTER}
 	err := tsv.StartService(target, dbcfgs)
